+++ test/unit/nft-api.test.ts
@@ -1766,4 +1766,7 @@ describe('NFT module', () => {
     it('handles upper case input', async () => {});
     it('handles errors', async () => {});
   });
+
+  // TODO: Add unit tests once we've implemented MockProvider.
+  describe('getMintedNfts()', () => {});
 });